requirements.txt
openai
speechrecognition
PyAudio
pyttsx3
python-dotenv
